\name{fev}
\alias{fev}
\docType{data}
\title{Forced Expiratory Volume (FEV) Data (fev)}
\description{
Sample of 654 youths, aged 3 to 19, in the area of East Boston
during middle to late 1970's. Interest concerns the relationship
between smoking and FEV. Since the study is necessarily
observational, statistical adjustment via regression models
clarifies the relationship.
}
\usage{data(fev)}
\format{
  A data frame with 654 observations on the following 5 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{fev}}{a numeric vector}
    \item{\code{ht}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{smoke}}{a numeric vector}
  }
}
\source{Kahn M (2005). An Exhalent Problem for Teaching Statistics. The Journal of Statistical Education, 13(2).

Rosner, B. (1999), Fundamentals of Biostatistics, 5th Ed., Pacific Grove, CA: Duxbury.
}
\keyword{datasets}
