\name{tmleNews}
\Rdversion{1.2.0-5}
\alias{tmleNews}
\title{Show the NEWS file (tmleNews)}
\description{
Shows recent changes and bug fixes documented in the \code{tmle} package NEWS file.}
\usage{
tmleNews(...)
}
\arguments{
  \item{...}{additional arguments passed to \code{RShowDoc}}
  }
\value{
  NONE
}

\author{Susan Gruber}

\seealso{
\code{\link{tmle}},
\code{\link{tmleMSM}}
}